Extract Platform into its own file
